Renamed node_fs_stats.ts to stats.ts
