+++ src/core/stats.ts
